requirements.txt
wget
numpy-stl>=2.17.1